+++ R/constructors.R
@@ -1,10 +1,68 @@
 
 ## User-visible constructors --------------------------------------------------
 
-## HAS_TESTS
-#' Create an Rvec
+#' Create a Blank Rvec
 #'
-#' Create an Object of Class `"rvec"`.
+#' Create an rvec, consisting entirely of `NAs`,
+#' with a given length and number of draws.
+#'
+#' The type of the object is taken from `x`.
+#' If `typeof(x)` is `"integer"`, for instance,
+#' then `new_rvec()` returns an object
+#' of class `"rvec_int"`.
+#'
+#' @param x Object with the intended type.
+#' Default is `double()`.
+#' @param length Desired length of rvec.
+#' Default is `0`.
+#' @param n_draw Number of draws of rvec.
+#' Default is `1000`.
+#'
+#' @returns An rvec.
+#'
+#' @seealso
+#' - [rvec()] [rvec_chr()], [rvec_dbl()], [rvec_int()], [rvec_lgl()]
+#'   Create an rvec from data.
+#' - [n_draw()] Query number of draws.
+#'
+#' @examples
+#' new_rvec()
+#' new_rvec(TRUE, length = 3, n_draw = 100)
+#'
+#' x <- new_rvec(length = 2)
+#' x[1] <- rnorm_rvec(n = 1, n_draw = 1000)
+#' x[2] <- runif_rvec(n = 1, n_draw = 1000)
+#' @export
+new_rvec <- function(x = double(), length = 0, n_draw = 1000) {
+  check_nonneg_num_scalar(length)
+  length <- as.integer(length)
+  check_nonneg_num_scalar(n_draw)
+  n_draw <- as.integer(n_draw)
+  type <- typeof(x)
+  if (type == "character")
+    na <- NA_character_
+  else if (type == "integer")
+    na <- NA_integer_
+  else if (type == "double")
+    na <- NA_real_
+  else if (type == "logical")
+    na <- NA
+  else {
+    valid_types <- c("character", "integer", "double", "logical")
+    cli::cli_abort(c("Invalid type.",
+                     i = "{.arg x} has type {.val {type}}.",
+                     i = "Valid types are: {.val {valid_types}}."))
+  }
+  m <- matrix(na, nrow = length, ncol = n_draw)
+  rvec(m)
+}
+
+
+## HAS_TESTS
+#' Create an Rvec from Data
+#'
+#' Create an object of class `"rvec"`, based
+#' on input data.
 #'
 #' Class `"rvec"` has four subclasses, each dealing with
 #' a diffent type:
@@ -31,7 +89,7 @@
 #' - an rvec.
 #'
 #' @param x A matrix, a list of vectors,
-#' or an atomic vector.
+#' an atomic vector, or an rvec.
 #'
 #' @returns
 #' An rvec with the following class:
@@ -43,6 +101,7 @@
 #' `"rvec_int"`, or `"rvec_lgl"`
 #'
 #' @seealso
+#' - [new_rvec()] Create a blank rvec.
 #' - [collapse_to_rvec()] Create rvecs within
 #'   a data frame.
 #' - [rnorm_rvec()], [rbinom_rvec()], etc. Create rvecs
@@ -300,20 +359,6 @@ rvec_lgl <- function(x = NULL) {
 
 
 ## Internal constructors ------------------------------------------------------
-
-## HAS_TESTS
-new_rvec <- function(data) {
-    if (is.character(data))
-        new_rvec_chr(data)
-    else if (is.double(data))
-        new_rvec_dbl(data)
-    else if (is.integer(data))
-        new_rvec_int(data)
-    else if (is.logical(data))
-        new_rvec_lgl(data)
-    else
-        cli::cli_abort("Internal error: {.arg data} has type {typeof(data)}.")
-}
 
 ## HAS_TESTS
 new_rvec_chr <- function(data) {